feat: move query to query.rs
