+++ contracts/ism-multisig/src/lib.rs
@@ -2,6 +2,7 @@ pub mod contract;
 mod error;
 pub mod event;
 pub mod execute;
+pub mod query;
 pub mod state;
 mod verify;
